# Make walk/prewalk calls static