--- a/lib/Parser.js
+++ b/lib/Parser.js
@@ -908,30 +908,30 @@ class Parser extends Tapable {
 
 	prewalkForInStatement(statement) {
 		if(statement.left.type === "VariableDeclaration")
-			this.prewalkStatement(statement.left);
+			this.prewalkVariableDeclaration(statement.left);
 		this.prewalkStatement(statement.body);
 	}
 
 	walkForInStatement(statement) {
 		if(statement.left.type === "VariableDeclaration")
-			this.walkStatement(statement.left);
+			this.walkVariableDeclaration(statement.left);
 		else
-			this.walkExpression(statement.left);
+			this.walkPattern(statement.left);
 		this.walkExpression(statement.right);
 		this.walkStatement(statement.body);
 	}
 
 	prewalkForOfStatement(statement) {
 		if(statement.left.type === "VariableDeclaration")
-			this.prewalkStatement(statement.left);
+			this.prewalkVariableDeclaration(statement.left);
 		this.prewalkStatement(statement.body);
 	}
 
 	walkForOfStatement(statement) {
 		if(statement.left.type === "VariableDeclaration")
-			this.walkStatement(statement.left);
+			this.walkVariableDeclaration(statement.left);
 		else
-			this.walkExpression(statement.left);
+			this.walkPattern(statement.left);
 		this.walkExpression(statement.right);
 		this.walkStatement(statement.body);
 	}