--- a/lib/Parser.js
+++ b/lib/Parser.js
@@ -1598,7 +1598,7 @@ class Parser extends Tapable {
 					this.scope.renames.set(param, null);
 					this.scope.definitions.add(param);
 				});
-			} else {
+			} else if(param) {
 				this.scope.renames.set(param, null);
 				this.scope.definitions.add(param);
 			}