--- a/test/cases/parsing/class/index.js
+++ b/test/cases/parsing/class/index.js
@@ -5,3 +5,8 @@ it("should parse classes", function() {
 	new A().a.should.be.eql("ok");
 	new B().a.should.be.eql("ok");
 });
+
+it("should parse methods", function() {
+	new X().b().should.be.eql("ok");
+	X.c().should.be.eql("ok");
+});