--- a/test/cases/parsing/class/module.js
+++ b/test/cases/parsing/class/module.js
@@ -14,6 +14,12 @@ export default class {
 	constructor() {
 		this.a = require("./a");
 	}
+	b() {
+		return require("./a");
+	}
+	static c() {
+		return require("./a");
+	}
 };
 
 export {