--- a/test/cases/parsing/pattern-in-for/module.js
+++ b/test/cases/parsing/pattern-in-for/module.js
@@ -1,0 +1,1 @@
+module.exports = "ok";