<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**What kind of change does this PR introduce?**

refactoring, perf

**Did you add tests for your changes?**

no

**If relevant, link to documentation update:**

n/a

**Summary**

Replace dynamic calls to `walk*` and `prewalk*` by static calls. Here below is the time spent by webpack building itself, prior and after the change:

Before:
```
❯ time node build_webpack.js
node build_webpack.js  7.75s user 0.54s system 115% cpu 7.189 total
```

After:
```
❯ time node build_webpack.js
node build_webpack.js  7.04s user 0.54s system 115% cpu 6.587 total
```

I did 10 runs before and 10 after. I took the median time. The best time before this change was never faster than the worst.

**Does this PR introduce a breaking change?**

no