@sokra 2017-12-30
Looks like a 10% performance improvement