@ooflorent 2017-12-30
I've written some basic tests cases. Since I'm not sure how to split `_(pre)walkNode`, I'll leave the PR as this for the moment. Feel free to contribute on my branch 😉