@ooflorent 2017-12-31
@sokra I've made the requested changes.